src/web/requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
requests==2.28.1
azure-monitor-opentelemetry
opentelemetry-instrumentation-flask
dotenv